Migrate mutual_links to appdir
